Move Credentials to config.go.
