+++ provider/gce/google/config.go
@@ -6,6 +6,7 @@ package google
 import (
 	"encoding/json"
 	"io"
+	"net/mail"
 
 	"github.com/juju/errors"
 
@@ -47,6 +48,46 @@ func ParseAuthFile(authFile io.Reader) (map[string]string, error) {
 		}
 	}
 	return data, nil
+}
+
+// Credentials holds the OAuth2 credentials needed to authenticate on GCE.
+type Credentials struct {
+	// ClientID is the GCE account's OAuth ID. It is part of the OAuth
+	// config used in the OAuth-wrapping network transport.
+	ClientID string
+
+	// ClientEmail is the email address associatd with the GCE account.
+	// It is used to generate a new OAuth token to use in the
+	// OAuth-wrapping network transport.
+	ClientEmail string
+
+	// PrivateKey is the private key that matches the public key
+	// associatd with the GCE account. It is used to generate a new
+	// OAuth token to use in the OAuth-wrapping network transport.
+	PrivateKey []byte
+}
+
+// Validate checks the credentialss for invalid values. If the values
+// are not valid, it returns errors.NotValid with the message set to
+// the corresponding OS environment variable name.
+//
+// To be considered valid, each of the credentials must be set to some
+// non-empty value. Furthermore, ClientEmail must be a proper email
+// address.
+func (gc Credentials) Validate() error {
+	if gc.ClientID == "" {
+		return NewInvalidCredential(OSEnvClientID, "", "missing ClientID")
+	}
+	if gc.ClientEmail == "" {
+		return NewInvalidCredential(OSEnvClientEmail, "", "missing ClientEmail")
+	}
+	if _, err := mail.ParseAddress(gc.ClientEmail); err != nil {
+		return NewInvalidCredential(OSEnvClientEmail, gc.ClientEmail, err)
+	}
+	if len(gc.PrivateKey) == 0 {
+		return NewInvalidCredential(OSEnvPrivateKey, "", "missing PrivateKey")
+	}
+	return nil
 }
 
 // ValidateConnection checks the connection's fields for invalid values.
